hice validaciones

+++ src/handlers/Favorites/deleteFavoriteHandler.js
@@ -4,6 +4,8 @@ const { deleteFavorite } = FavoritesControllers;
 const addFavoriteHandler = async (req, res) => {
     try {
         const { idUser, idProduct } = req.body;
+        if (!idUser || !idProduct) return res.status(400).json({ error: "faltan datos" });
+
         const result = await deleteFavorite(idUser, idProduct);
         return res.status(200).json(result);
     } catch (error) {
